классы сервера и клиента распределены в разные файлы
