+++ src/client/test_client.cpp
@@ -3,73 +3,9 @@
 #include <thread>
 #include <atomic>
 
-using boost::asio::ip::tcp;
+#include "Client.hpp"
 
-class Client
-{
-private:
-    tcp::socket socket;
-    std::string read_buffer;
 
-    void read()
-    {
-        boost::asio::async_read_until(socket, boost::asio::dynamic_buffer(read_buffer), '\n',
-            [this](boost::system::error_code ec, std::size_t length)
-            {
-                if (!ec)
-                {
-                    std::cout << "Received: " << read_buffer.substr(0, length);
-                    read_buffer.erase(0, length);
-                    read(); // Continue reading
-                }
-                else
-                {
-                    std::cerr << "Read error: " << ec.message() << "\n";
-                    socket.close();
-                }
-            });
-    }
-
-    void write()
-    {
-        boost::asio::post([this]()
-        {
-            std::string write_buffer;
-            std::getline(std::cin, write_buffer);
-            
-            write_buffer += '\n'; // Add delimiter
-            
-            boost::system::error_code ec;
-            boost::asio::async_write(socket, boost::asio::buffer(write_buffer), 
-                [this](boost::system::error_code ec, std::size_t length)
-                {
-                    if (!ec)
-                        write();
-                    else
-                    {
-                        std::cerr << "Write error: " << ec.message() << "\n";
-                        socket.close();
-                    }
-                });
-        });
-    }
-
-public:
-    Client(boost::asio::io_context& io_context, tcp::socket&& sock)
-        : socket(std::move(sock))
-    {
-        // Запускаем чтение в основном потоке io_context
-        read();
-        
-        // Запускаем запись в отдельном потоке
-        write();
-    }
-
-    ~Client()
-    {
-        socket.close();
-    }
-};
 
 int main() 
 {
